pyproject.toml
[project]
name = "mkdocs-mike-oauth"
version = "0.1.0"
description = ""
authors = [
    {name = "Gabriel Dornas",email = "gabrielbdornas@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "mkdocs-material (>=9.6.9,<10.0.0)",
    "taskipy (>=1.14.1,<2.0.0)"
]

[tool.poetry.group.dev.dependencies]
ipdb = "^0.13.13"
pytest = "^8.3.5"

[tool.taskipy.tasks]
serve = { cmd = "mkdocs serve", help = "Starts server." }
build = { cmd = "mkdocs build", help = "Deploy static files to gh-pages." }
docker_build = { cmd = "docker build --no-cache -t my-mkdocs-site .", help = "Builds docker image locally." }
docker_run = { cmd = "docker run --rm -p 8080:4180 my-mkdocs-site", help = "Runs docker container locally." }

[tool.poetry]
package-mode = false

[tool.poetry.plugins."mkdocs.plugins"]
"exclude-pages" = "plugins.exclude_pages:ExcludePagesPlugin"
